[mod] refactor to use domain by specifying
